% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvFolds.R
\name{cvFolds}
\alias{cvFolds}
\title{Creates cross-validation folds}
\usage{
cvFolds(
  data = train,
  foldNumber = 5,
  stratifyOnVar = FALSE,
  whatVarToStratifyOn = "var"
)
}
\arguments{
\item{data}{The training data set}

\item{foldNumber}{the number of folds for cross-validation. Default is 5.}

\item{stratifyOnVar}{Logical. Should the folds be stratified based on the response? If so, set TRUE.}

\item{whatVarToStratifyOn}{Character. What is the name of the variable to stratify on?}
}
\value{

}
\description{
Create folds based on training data fed in.
}
