% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialTimeSeriesEval.R
\name{initialTimeSeriesEval}
\alias{initialTimeSeriesEval}
\title{Initial Time Series Evaluation}
\usage{
initialTimeSeriesEval(df = dts, response = resp, predictor = pred)
}
\arguments{
\item{df}{The time series formatted data frame. Columns must be named "Date" and "Value" respectively.}

\item{response}{The variable that is the response for analysis.}

\item{predictor}{The variable that is the predictor.}
}
\value{

}
\description{
Gives overview of time series object. Plots time series, seasonality, and acf.
}
