% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearRegress.R
\name{linearRegress}
\alias{linearRegress}
\title{Linear Regression.}
\usage{
linearRegress(
  response = response,
  computeMarginalEffects = FALSE,
  data = df,
  train = train_df,
  test = test_df,
  tidyModelVersion = FALSE,
  recipe = rec,
  folds = folds,
  evalMetric = "rmse"
)
}
\arguments{
\item{response}{The variable that is the response for analysis.}

\item{computeMarginalEffects}{Logical. Compute marginal effects for lm model?}

\item{data}{The entire data frame.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{tidyModelVersion}{Logical. Run a tidymodel version of linear regression?}

\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{evalMetric}{The regression metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Runs a linear regression model, and either:
\enumerate{
\item fits a basic lm() model and shows diagnostics and model fit
\item evaluates it on training and testing set, and tunes hyperparameters.
}
}
