% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRegBinary.R
\name{logRegBinary}
\alias{logRegBinary}
\title{Logistic Regression.}
\usage{
logRegBinary(
  recipe = rec,
  folds = cvFolds,
  train = datTrain,
  test = datTest,
  response = response,
  evalMetric = "bal_accuracy"
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{response}{The variable that is the response for analysis.}

\item{evalMetric}{The classification metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Runs a logistic regression model, evaluates it on training and testing set, and tunes hyperparameters.
}
