% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logRegMulti.R
\name{logRegMulti}
\alias{logRegMulti}
\title{Multinomial Regression.}
\usage{
logRegMulti(
  recipe = rec,
  folds = cvFolds,
  train = train_df,
  test = test_df,
  response = response,
  levelNum = 15,
  evalMetric = "bal_accuracy"
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{response}{The variable that is the response for analysis.}

\item{levelNum}{Numeric. The number of levels you want the grid to search on. Default is 15.}

\item{evalMetric}{The classification metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Runs a multinomial regression model, evaluates it on training and testing set, and tunes hyperparameters.
}
