% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTimeSeries.R
\name{modelTimeSeries}
\alias{modelTimeSeries}
\title{Model Time Series}
\usage{
modelTimeSeries(
  train = tr,
  test = ts,
  df = dts,
  response = resp,
  predictor = pred,
  folds = cvFolds,
  days = 7
)
}
\arguments{
\item{train}{Training data set.}

\item{test}{The testing data set.}

\item{df}{The time series formatted data frame. Columns must be named "Date" and "Value" respectively.}

\item{response}{The variable that is the response for analysis.}

\item{predictor}{The variable that is the predictor.}

\item{folds}{A rsample::vfolds_cv object.}

\item{days}{The numbers of days desired to forecast for. Default is 7.}
}
\value{

}
\description{
Runs many time series models as well as combinations of the time series.
}
