% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleTimeSeries.R
\name{resampleTimeSeries}
\alias{resampleTimeSeries}
\title{Resample Time Series}
\usage{
resampleTimeSeries(df = dts, table = output$forecastTable, sliceLimit = 5)
}
\arguments{
\item{df}{The time series formatted data frame. Columns must be named "Date" and "Value" respectively.}

\item{table}{A modeltime table object.}

\item{sliceLimit}{The number of slices to have in resample. Default is 5.}
}
\value{

}
\description{
Wrapper for modeltime.resample. Resamples time series to see how they compare over different time chunks.
}
