% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmClassif.R
\name{svmClassif}
\alias{svmClassif}
\title{Support Vector Machine Classification}
\usage{
svmClassif(
  response = response,
  recipe = rec,
  folds = folds,
  train = train_df,
  test = test_df,
  gridNumber = 15,
  evalMetric = "rmse"
)
}
\arguments{
\item{response}{The variable that is the response for analysis.}

\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{gridNumber}{The size of the grid to tune on.}

\item{evalMetric}{The regression metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Fits a radial basis Support Vector Machine Classification Model.
}
