% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainTestSplit.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Split data into training and testing}
\usage{
trainTestSplit(
  data = df,
  splitAmt = 0.8,
  timeDependent = FALSE,
  responseVar = "nameOfResponseVar",
  stratifyOnResponse = FALSE
)
}
\arguments{
\item{data}{The data set of interest.}

\item{splitAmt}{The amount of data you want in the training set. Default is .8}

\item{timeDependent}{Logical. Is your data time-dependent? If so, set TRUE.}

\item{responseVar}{Name of response variable in analysis.}

\item{stratifyOnResponse}{Logical. Should the training and testing splits be stratified based on the response? If so, set TRUE.}
}
\value{

}
\description{
Create a training and testing data set. Also returns a bootstrapped version of the training data set.
}
