% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgBinaryClassif.R
\name{xgBinaryClassif}
\alias{xgBinaryClassif}
\title{XGBoost Binary Classification}
\usage{
xgBinaryClassif(
  gridNumber = 10,
  recipe = rec,
  folds = cvFolds,
  train = datTrain,
  test = datTest,
  response = response,
  treeNum = 100,
  calcFeatImp = TRUE,
  evalMetric = "bal_accuracy"
)
}
\arguments{
\item{gridNumber}{Numeric. Size of the grid you want XGBoost to explore. Default is 10.}

\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{response}{The variable that is the response for analysis.}

\item{treeNum}{The number of trees to evaluate your model with.}

\item{calcFeatImp}{Do you want to calculate feature importance for your model? If not, set = FALSE.}

\item{evalMetric}{The classification metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Runs XGBoost for binary classification.
}
