% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgRegress.R
\name{xgRegress}
\alias{xgRegress}
\title{XGBoost Regression}
\usage{
xgRegress(
  gridNumber = 10,
  recipe = rec,
  folds = cvFolds,
  train = train_df,
  test = test_df,
  response = response,
  treeNum = 100,
  calcFeatImp = TRUE,
  evalMetric = "rmse"
)
}
\arguments{
\item{gridNumber}{Numeric. Size of the grid you want XGBoost to explore. Default is 10.}

\item{recipe}{A recipe object.}

\item{folds}{A rsample::vfolds_cv object.}

\item{train}{The training data set.}

\item{test}{The testing data set.}

\item{response}{The variable that is the response for analysis.}

\item{treeNum}{The number of trees to evaluate your model with.}

\item{calcFeatImp}{Do you want to calculate feature importance for your model? If not, set = FALSE.}

\item{evalMetric}{The regression metric you want to evaluate the model's accuracy on.}
}
\value{

}
\description{
Runs XGBoost for regression
}
