% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRecipe.R
\name{createRecipe}
\alias{createRecipe}
\title{Creates recipes}
\usage{
createRecipe(data = data, responseVar = "response")
}
\arguments{
\item{data}{The training data set}

\item{responseVar}{the variable that is the response for analysis.}
}
\value{

}
\description{
Create recipes
}
